refactor(legacy-confirm-modal): rename
