+++ packages/bezier-react/src/components/Modals/LegacyConfirmModal/ConfirmModal.stories.tsx
