+++ packages/bezier-react/src/components/Modals/LegacyConfirmModal/ConfirmModal.types.tsx
